Addition and Subtraction

+++ src/com/mycompany/basicmathoperations/simple/Multiply.java
@@ -9,10 +9,10 @@ package com.mycompany.basicmathoperations.simple;
  */
 public class Multiply {
 	/**
-	 * To multiply 2 int numbers
+	 * To multiply 2 float numbers
 	 * @return num1 * num2
 	 */
-	public int multiply(int num1, int num2) {
+	public float multiply(float num1, float num2) {
 		return num1 * num2;
 	}
 }
